Check for a shutdown signal in the webjob program before initializing data
We have seen intermittent cases where a couple of devices would be
bootstrapped, then about 7 seconds later four more devices. After some
investigation we found that on every fresh cloud deploy the web job would
go through two starts -- we suspect this is as it goes through the
staging, warmup and swap to production.

To avoid letting this happen we will watch the location of the shutdown
file that is created when shutdown is scheduled. If we see that before we
enter the initialization method then we will skip initialization since it
will likely be aborted mid-process.

In Debugging we have sometimes seen cases where initialization is started,
then the shutdown signal is immediatly received. However, in these cases
initialization has been able to complete successfully.

It may not be a perfect solution, but for now I think it's pretty solid.