Add a timer on the data initialization to improve chances of seeing the shutdown
The shutdown command comes pretty reliably within about 7 seconds if it is
going to come. So we will wait 10 seconds before doing the data
initialization to improve the chances we do that initialization on the
second boot, which will remain booted and furthre reduce the chances of an
aborted partial initialization.