Fix how we initialize default min/max lat/long when gathering device location data
We were not allowing the map bounds to initialize correctly because we
were initializing around Seattle. Any maxLattitude that was not greater
than Seattle's lattitude was ignored. Any minLongitude that is greater
than Seattle's longitude was being ignored

If we initialize max<x> to double.min, we guarantee that any device
locations will be honored on the first loop. Conversely for min<x>. if the
list ends up being empty, then we will re-initialize the min/max lat/long
to center around Seattle.