Update DocDbUtil to use a query when getting the db or collection
This avoids errors that can be thrown on the server when other collections
exist that were added through the portal. When that is the case, the new
collection was created with a newer api than we are using to retrieve. Using a
simple request to retrieve the list of databases or collections causes the
server to reject the request because it can't read that new collection.
However, querying specifically for the collection we want avoids an
attempt to read the new collection and avoids the error.

Also, pulled out the bulk of the query logic into an internal method to be
resued by the three places we're calling it.