Unify all timestamp encoding for DocDb calls
We were previously getting the timestamp in two different places, one for
the date header and one for the authorization header. Now we create the
formatted timestamp string in one place and place both headers on the call
immediately before making the call. This reduces the chance for rejected
docDb calls because of lapsed time.