Add two buttons on the device list
In this code change, we are adding two buttons in the dataTable. It was implemented by add an empty div with the built-in property 'dom' of dataTable, then create two buttoned dynamically.

Reminder: The actions behind the two buttons are not there. Currently, they could only toggle the device detail panel as a placeholder.