Publish the C simulator
The C simulator acts as a workaround for the E2E test before the C# client SDK get ready. It should be removed once the official SDK and simulator were updated.

By this code change, we will put it into the whole build process to keep the binary always up-to-date. One more change in the C simulator is it will check for the existence of parent process (the simulator web job) and exit when the parent process was closed. So there is no need to terminate the C simulators while debugging.

All the header files and library were come from the preview SDK repository.