Align with IoT Hub service SDK (Microsoft.Azure.Devices.1.1.0-preview-005)
Microsoft.Azure.Devices.1.1.0-preview-005 introduced a breaking change: the items in the TwinCollection use a nested TwinCollection object rather JContainer for hierarchy data structure. It happens only for desired and reported properties, but not for tag.

Since the declared type in the SDK is always 'object'. This change was not cause a compiler error. But it do affected to the flatten property list we showing on the RM portal. So we have to align with this new implementation.