Retrieve LastUpdatedUtc from Twin
In this code change, we will retrieve the last updated time from twin. It is a backend only change. The UI change will be submitted in other CR.

Currently, LastUpdateUtc is not a "stable" property of the Twin. For some desired/reported properties, it is not available since the property was not extracted as the TwinCollectionValue. So we just retreat it as optional.