Show name rather than ID of Job on the device detail panel
A new Storage Table (named as JobList by default) introduced to store JobID, QueryName and JobName.

Considering there will be no job in the new added list, since the "Schedule job" feature is not ready, we added a mock data generator to collect actual job history to feed the table. It was disabled by default. To enable it, please uncomment line 29 in Global.asax.cs. Reminder: it will remove all the content in table QueryList and JobTable.