[BugFix] Exception raised due to empty job name
Job name was retrieved from Azure storage table. It could be uninitialized or incorrectly modified as empty string. In that case, an exception will be raised while we trying to use this empty string as the text of hyperlink. By this code change, we will check if the text is empty, and replace it with the "[jobId]" if it is the case.