Upgrade IoT Hub service SDK to 1.1.0 (official version)
IoT Hub DM feature was GA yesterday. And the NuGet package of the service SDK was updated on last Thursday. We are going to upgrade the package in our project to official version.

There is one minor change in interface: the returned value type of C2D method was changed from string to int.