Add support to get/set value of tag/property at any level
In previous implementation, there is a limitation in the Get/Set method of TwinExtension: only the value at top level could be get/set. In other words, Twin.Tags.Get("Location.City") is not supported (it will cause exception raised in SDK). In this code change, we are going to enable get/set at any level.