Improve performance for device query
Currently, device metadata was stored in two place - DocumentDB (for command, telemetries, etc.) and IoT Hub (for twin). We must try to merge those two data source for each query. In previous code, we query Document DB separately for each devices we retrieved from the IoT Hub. It may cause redundant initializing effort.

In this code change, we use one single query for all the devices to improve the performance. Furthermore, we are going to delay the query on Document DB until the result was paginated. It will bring some benefit for pagination case.