Add job indicators on dashboard
Considering we are going to enable customized dashboard, the job indicators were not hardcoded on the view. The front-end will query for "definitions" at the beginning, create elements according to the definition, then try to query for values. Currently, the definitions were hard coded in back-end code. Later, we will try to move it into user settings or some other storage.