Retrieve the device job history in asynchronizing style
In this code change, we are going to move the operation "retrieve device jobs" to a separated HTTP request, to speed-up showing of the device detail panel. The device job list was returned as a partial view which was rendered by back-end, to avoid lengthy code parsing the job list and building HTML elements at the front-end.