Remove the workaround template explicitly declaring for DM feature
In previous weeks, an explicitly declaration for DM feature ("features": "DeviceManagement") was required to enable DM feature of IoT Hub. We added a workaround template to help team to deployment test environment. Considering the DM feature of IoT Hub was already GA, and we verified a deployment template without explicitly declaration for DM feature could create an IoT Hub with DM feature, it is not necessary now. So we are going to remove this workaround template.