[BugFix] Device location was not retrieved from Twin
Device location (latitude and longitude) was stored in DocumentDB together with all the device "properties" in original version. In this branch, we planed to switch to IoTHub Twin. The data source of most of UI (such as device list and device detail panel) was switched at the same time. While the data source of the device location showing on the dashboard was not changed. In this code change, we are going to fix this error.