Use existing device properties rather than regenerate them every time simulator rebooted
In previous code, the simulator will regenerate all the device properties (such as installedMemory, Location and Platform) every time it rebooted. It will confuse user especially when they trying to update firmware. By this code change, we will try to retrieve the existing properties from IoTHub reported properties, and only use the re-generated values when the reported value was absent.