Select IoTHub transport type according to the run time environment
Days ago, we found that "Mqtt" is not a valid transport type for IoT Hub device simulator running in WebJob. So we changed the transport to Mqtt_WebSocket_Only. While we just noticed that Mqtt_WebSocket_Only will introduce tons of "Channel closed" exception while the device simulator running in local debug environment. So we have to select transport according to the environment:

If there is no environment variable named as "WEBSITE_HOSTNAME", we suppose the device simulator is running in local debug environment, the transport type will be Mqtt. Otherwise, the default value Mqtt_WebSocket_Only will be chosen.