Shorten auto-generated device id
Currently, the auto-generated device id is just a GUID. To optimize the visual effect, we are going to shorten it as pattern "SampleDevice_nnnn".

Considering the cost of holding a random generator across HTTP requests, the "nnnn" part will be generated from the first 4 bytes of a GUID, and verified to ensure no conflict.