Pick version from downloaded firmware file rather than URL itself
In previous code inherited from the C simulator, we just try to get the firmware version from the firmware URL itself, rather than try to download the file through internet. It cause difficult to demo some scenarios, such as downloading failed. By this code change, we are going to download the file and pick the version from file. Currently, the "firmware" file could only contains one line. The whole content will be treated as the version.

On the other hand, apply and reboot are still mocked. So a firmware with content "applyFail" will cause an applyFail error, to demo some error scenario.