Add workaround IoTHubTransport wrapper
There is an issue in the IoT Hub device SDK which will cause random exception while trying to update reported properties. IoT Hub team suggested us to create a new DeviceClient object as workaround.

Since the exception will be raised randomly, in rare case it will even happen in the new created DeviceClient. So it is not enough to just create the new DeviceClient and try one time. In this code change, we are going to add a new IoTHubTransport wrapper to buffer the out-going reported properties until it was successfully updated to IoT Hub. Considering the telemetry event will be generated periodically, we will just drop those event if it was failed to be sent.

To avoid resource leak, we dispose the current DeviceClient when we are going to create a new one. But it may cause null referenced exception under some racing condition. So we are going to move all the operations on the DeviceClient to a single task “SendLoop” - We introduced two buffer for the out-going reported properties and telemetry events. All the "Send" operation on the IIoTHubTransport were converted to a "enqueue" operation and returned immediately. The “SendLoop” task will poll the buffer periodically and send them to IoT Hub. If any exception raised while sending to IoT Hub, the “SendLoop” task will mark the DeviceClient as unavailable, then dispose it and re-create a new one in next round.